requirements.txt
discord.py==2.0.1
python-dotenv==0.20.0
requests
urllib3==1.26.6